+++ micro-judge/src/io/xread.rs
@@ -1,8 +1,8 @@
 use super::conn_pool::Pool;
-use super::entry_id::AllEntryIds;
 use super::topics::*;
 use super::DeserError;
 use crate::model::*;
+use crate::repo::entry_id::AllEntryIds;
 use redis;
 use std::collections::HashMap;
 use std::str::FromStr;
