+++ micro-judge/src/repo/entry_id.rs
@@ -1,7 +1,8 @@
-use super::conn_pool::Pool;
-use super::redis_keys::*;
-use super::xread::XReadEntryId;
-use super::FetchErr;
+use crate::io;
+use io::conn_pool::Pool;
+use io::redis_keys::*;
+use io::xread::XReadEntryId;
+use io::FetchErr;
 use r2d2_redis::redis;
 use redis::Commands;
 use std::collections::HashMap;
